+++ src/components/certificates/certificates-preview.tsx
@@ -1,6 +1,6 @@
 import { useStore } from "./hooks";
 import { MonthsPreview } from "./months/months-preview";
-import { YearsPreview } from "./years";
+import { YearsPreview } from "./years/years";
 
 export function CertificatesPreview() {
   const certificate = useStore((state) => state.certificate);
