+++ src/components/certificates/certificates.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, useState } from "react";
 import { Months } from "./months/months";
-import { Years } from "./years";
+import { Years } from "./years/years";
 import { CheckBox } from "./check-box";
 import { useForm } from "react-hook-form";
 import { Details } from "../details/details";
